requirement.txt
sklearn
numpy
matplotlib
scipy
numpy
tensorflow
pytorch
fastai